[ANDROSDK-1956] Move user fields to network
