+++ core/src/main/java/org/hisp/dhis/android/core/user/UserObjectRepository.kt
@@ -32,9 +32,9 @@ import org.hisp.dhis.android.core.arch.repositories.children.internal.ChildrenAp
 import org.hisp.dhis.android.core.arch.repositories.`object`.internal.ObjectRepositoryFactory
 import org.hisp.dhis.android.core.arch.repositories.`object`.internal.ReadOnlyOneObjectRepositoryImpl
 import org.hisp.dhis.android.core.arch.repositories.scope.RepositoryScope
-import org.hisp.dhis.android.core.user.internal.UserFields
 import org.hisp.dhis.android.core.user.internal.UserRoleChildrenAppender
 import org.hisp.dhis.android.core.user.internal.UserStore
+import org.hisp.dhis.android.network.user.UserFields
 import org.koin.core.annotation.Singleton
 
 @Singleton(binds = [UserObjectRepository::class])
